requirements.txt
Faker==23.3.0
mysql-connector==2.2.9
mysql-connector-python
pandas